move BISON intermediate values into subclass
